The crash fix for 147044 has been reported to work, committing it. BUG: 147044
svn path=/branches/koffice/1.6/koffice/; revision=725851

+++ lib/kotext/KoTextObject.cpp
@@ -303,11 +303,12 @@ void KoTextObject::UndoRedoInfo::clear()
                 KoTextDocCommand * cmd = textobj->deleteTextCommand( textdoc, id, index, text.rawData(), customItemsMap, oldParagLayouts );
                 textdoc->addCommand( cmd );
                 Q_ASSERT( placeHolderCmd );
-                placeHolderCmd->addCommand( new KoTextCommand( textobj, /*cmd, */QString::null ) );
-                // Deleting any custom items -> let them add their command
-                if ( !customItemsMap.isEmpty() )
-                {
-                    customItemsMap.deleteAll( placeHolderCmd );
+                if (placeHolderCmd) {
+                    placeHolderCmd->addCommand( new KoTextCommand( textobj, /*cmd, */QString::null ) );
+                    // Deleting any custom items -> let them add their command
+                    if ( !customItemsMap.isEmpty() ) {
+                        customItemsMap.deleteAll( placeHolderCmd );
+                    }
                 }
            } break;
             case Invalid:
